Move VCM main to src
