+++ src/general/classes/UserInputs.py
@@ -1,4 +1,4 @@
-from src.general.classes.PlatformInfo import PlatformInfo
+from general.classes.PlatformInfo import PlatformInfo
 
 
 class UserInputs:
